requirements.txt
# Essenciais para o dashboard
streamlit==1.48.1
pandas==2.3.1
numpy==2.3.2
openpyxl==3.1.5
pyarrow==21.0.0
altair==5.5.0
requests==2.32.5
agno
openai>=1.30.0
python-dotenv
